Got rid of BaseOpMode and BaseAutonomous because the FTC Robot controller doesn't support multiple extends
