--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/BaseOpMode.java
@@ -1,52 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.hardware.lynx.LynxI2cColorRangeSensor;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
-
-/**
- * Created by serge on 12/6/2017.
- * Base class for all OpModes
- */
-public abstract class BaseOpMode extends LinearOpMode {
-
-    // Motors
-    protected DcMotor motor0;
-    protected DcMotor motor1;
-    protected DcMotor spoolMotor;
-
-    // Servos
-    protected Servo servo0;
-    protected Servo servo1;
-    protected Servo servo2;
-    protected Servo servo3;
-    protected Servo servo5;
-
-    // Color sensor
-    protected LynxI2cColorRangeSensor color0;
-
-    /**
-    * Initialize the variables for the OpMode
-    */
-    protected void initOpMode() {
-        // Motors initialization
-        motor0 = hardwareMap.get(DcMotor.class, "motor0");
-        motor1 = hardwareMap.get(DcMotor.class, "motor1");
-        motor0.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        motor1.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        spoolMotor = hardwareMap.get(DcMotor.class, "spoolMotor");
-        // Servos initialization
-        servo0 = hardwareMap.get(Servo.class, "servo0");
-        servo1 = hardwareMap.get(Servo.class, "servo1");
-        servo2 = hardwareMap.get(Servo.class, "servo2");
-        servo3 = hardwareMap.get(Servo.class, "servo3");
-        servo5 = hardwareMap.get(Servo.class, "servo5");
-        // Sensors intialization
-        color0 = hardwareMap.get(LynxI2cColorRangeSensor.class, "color0");
-
-        // initialize the spool encoder
-        spoolMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        spoolMotor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-    }
-}
