+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,7 +1,6 @@
-#Thu Jun 01 08:33:11 PDT 2017
+#Wed Dec 06 19:49:12 PST 2017
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.5-all.zip
-
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.1-all.zip
